+++ packages/insomnia-inso/src/examples/script-add-header.yml
@@ -1,0 +1,44 @@
+_type: export
+__export_format: 4
+__export_date: 2021-12-03T07:42:43.250Z
+__export_source: insomnia.desktop.app:v2021.6.0
+resources:
+  - _id: req_178f98ccc6b240f5a1f67c492b6bf9a6
+    parentId: wrk_5b5ab67830944ffcbec47528366ef403
+    modified: 1636142586648
+    created: 1636141100570
+    url: http://127.0.0.1:4010/basic/auth
+    name: basic auth by pre request script
+    description: ""
+    method: GET
+    body: {}
+    parameters: []
+    preRequestScript: "insomnia.request.addHeader({key: 'Authorization', value: 'Basic dXNlcjpwYXNz' });"
+    metaSortKey: -1636141100570
+    isPrivate: false
+    settingStoreCookies: true
+    settingSendCookies: true
+    settingDisableRenderRequestBody: false
+    settingEncodeUrl: true
+    settingRebuildPath: true
+    settingFollowRedirects: global
+    _type: request
+  - _id: wrk_5b5ab67830944ffcbec47528366ef403
+    parentId: null
+    modified: 1636140994423
+    created: 1636140994423
+    name: Minimal Collection 2
+    description: ""
+    scope: collection
+    _type: workspace
+  - _id: env_9e44ac71d3a51fc6a8951b8968e45c44c5236355
+    parentId: wrk_5b5ab67830944ffcbec47528366ef403
+    modified: 1720533288414
+    created: 1720533288414
+    name: Base Environment
+    data: {}
+    dataPropertyOrder: null
+    color: null
+    isPrivate: false
+    metaSortKey: 1720533288414
+    _type: environment
